+++ blocklets/ai-runtime/api/src/libs/resource.ts
@@ -261,14 +261,21 @@ const loadResourceKnowledge = async () => {
     .filter((i): i is NonNullable<typeof i> => !!i);
 };
 
-async function reloadResources() {
+async function initResources({ reload }: { reload?: boolean } = {}) {
+  if (reload) cache.promise = undefined;
   cache.promise ??= loadResources();
   return cache.promise;
 }
 
-export const getResourceProjects = async (type: ResourceType) => {
-  const resources = await reloadResources();
-  return resources.agents[type]?.projects ?? [];
+export const getResourceProjects = async ({ blockletDid, type }: { blockletDid?: string; type: ResourceType }) => {
+  const resources = await initResources();
+  const list = resources.agents[type]?.projects ?? [];
+
+  if (blockletDid) {
+    return list.filter((i) => i.blocklet.did === blockletDid);
+  }
+
+  return list;
 };
 
 export interface ResourceKnowledge {
@@ -281,7 +288,7 @@ export interface ResourceKnowledge {
 }
 
 export const getResourceKnowledgeList = async () => {
-  const resources = await reloadResources();
+  const resources = await initResources();
   return resources.knowledge.knowledgeList;
 };
 
@@ -292,23 +299,8 @@ export const getResourceKnowledgeWithData = async ({
   blockletDid: string;
   knowledgeId: string;
 }) => {
-  const resources = await reloadResources();
+  const resources = await initResources();
   return resources.knowledge.blockletMap[blockletDid]?.knowledgeMap[knowledgeId];
-};
-
-export const getResourceBlockletFromResource = async ({
-  blockletDid,
-  type,
-}: {
-  blockletDid: string;
-  type?: ResourceType | ResourceType[];
-}) => {
-  const resources = await reloadResources();
-  for (const t of type ? [type].flat() : ResourceTypes) {
-    const p = resources.agents[t]?.blockletMap[blockletDid];
-    if (p) return p;
-  }
-  return undefined;
 };
 
 export const getProjectFromResource = async ({
@@ -320,10 +312,10 @@ export const getProjectFromResource = async ({
   projectId: string;
   type?: ResourceType | ResourceType[];
 }) => {
-  const resources = await reloadResources();
+  const resources = await initResources();
   for (const t of type ? [type].flat() : ResourceTypes) {
     const p = resources.agents[t]?.blockletMap[blockletDid]?.projectMap[projectId];
-    if (p) return p.project;
+    if (p) return p;
   }
   return undefined;
 };
@@ -337,7 +329,7 @@ export const getMemoryVariablesFromResource = async ({
   projectId: string;
   type?: ResourceType | ResourceType[];
 }) => {
-  const resources = await reloadResources();
+  const resources = await initResources();
   for (const t of type ? [type].flat() : ResourceTypes) {
     const p = resources.agents[t]?.blockletMap[blockletDid]?.projectMap[projectId];
     if (p) return p.memory?.variables;
@@ -356,7 +348,7 @@ export const getAssistantFromResourceBlocklet = async ({
   agentId: string;
   type?: ResourceType | ResourceType[];
 }) => {
-  const resources = await reloadResources();
+  const resources = await initResources();
   for (const t of [type].flat()) {
     const blocklet = resources.agents[t]?.blockletMap[blockletDid];
     const project = blocklet?.projectMap[projectId];
@@ -369,11 +361,17 @@ export const getAssistantFromResourceBlocklet = async ({
 
 export function initResourceStates() {
   async function reloadStatesWithCatch() {
-    cache.promise = undefined;
-
-    await reloadResources().catch((error) => {
-      logger.error('load resource states error', { error });
-    });
+    logger.info('reload resource states');
+    await initResources({ reload: true })
+      .then((resource) => {
+        logger.info('reload resource states success', {
+          projects: Object.values(resource.agents).reduce((res, i) => i.projects.length + res, 0),
+          knowledge: resource.knowledge.knowledgeList.length,
+        });
+      })
+      .catch((error) => {
+        logger.error('reload resource states error', { error });
+      });
   }
 
   reloadStatesWithCatch();