+++ blocklets/ai-runtime/src/libs/constants.ts
@@ -1,1 +1,0 @@
-export const AIGNE_RUNTIME_CUSTOM_COMPONENT_ID = 'grc9q1cveub6pnl8';