+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,69 @@
+## 0.3.1 (2024-6-22)
+
+- fix: allow select components from pages-kit and resource blocklets
+
+## 0.3.0 (2024-6-22)
+
+- chore: update deps
+
+## 0.2.66 (2024-6-21)
+
+- fix: invalid knowledge resource blocklet metadata
+
+## 0.2.65 (2024-6-21)
+
+- fix: generated images not display in the debug view
+
+## 0.2.64 (2024-6-21)
+
+- chore: add screenshots for ai-runtime
+- chore: hide navigation for aigne runtime and generated resource blocklets
+
+## 0.2.63 (2024-6-21)
+
+- fix: reset loading indicator after saving dialog is canceled
+- chore: set agent detail link to the agent blocklet mount point
+- chore: display agent logo and name in the preview header
+
+## 0.2.62 (2024-6-21)
+
+- chore: improve projects/knowledge page ux
+
+## 0.2.61 (2024-6-21)
+
+- chore: display icon before the custom input/output
+- fix: the logo flashes when opening the agent view
+- chore: add docLink and placeholder for secret inputs
+- fix: run agent error if there is a input to retrieve chat history
+
+## 0.2.60 (2024-6-20)
+
+- feat: add home page to display agent list
+- chore: update logo and readme
+
+## 0.2.59 (2024-6-20)
+
+- fix: publish resource blocklet error in strict mode
+
+## 0.2.58 (2024-6-20)
+
+- fix: display loading indicator for publish button
+- fix: set first agent as entry
+- chore: add blocklet isolation mode support
+
+## 0.2.57 (2024-6-19)
+
+- chore: update deps
+
+## 0.2.56 (2024-6-19)
+
+- fix: inherit identity from parent agent
+
+## 0.2.55 (2024-6-19)
+
+- fix: display knowledge idle status
+- fix: throw error if the searched documents too large
+
 ## 0.2.54 (2024-6-19)
 
 - feat: allow configuring the attachment of the URL and inputs in the shared content